refactor(aggregate): move aggregate-specific naming and event application out of kernel
This keeps @redemeine/kernel focused on shared contracts while colocating aggregate runtime behavior and aggregate-only helper types with @redemeine/aggregate.
