+++ packages/aggregate/src/createAggregate.ts
@@ -1,4 +1,4 @@
-import { Event, Command, EventEmitterFactory, EventType, CommandType, NamingStrategy, SelectorsMap, AggregateHooks, MapCommandsToPayloads, PluginContext, PluginExtensions, CommandContext, CommandIntents, MergePluginExtensions, RedemeinePlugin, ReadonlyDeep, Merge, AllKeys, applyEvent, defaultNamingStrategy } from '@redemeine/kernel';
+import { Event, Command, EventEmitterFactory, EventType, CommandType, NamingStrategy, SelectorsMap, AggregateHooks, MapCommandsToPayloads, PluginContext, PluginExtensions, CommandContext, CommandIntents, MergePluginExtensions, RedemeinePlugin, ReadonlyDeep } from '@redemeine/kernel';
 import { MixinPackage } from './createMixin';
 import { EntityPackage } from './createEntity';
 import { createCommandProcessor } from './createCommandProcessor';
@@ -7,6 +7,10 @@ import { createCommandCreatorsProxy } from './proxies/createCommandCreatorsProxy
 import { createCommandContextProxy } from './proxies/createCommandContextProxy';
 import { RedemeineCommandDefinition, RedemeineEventDefinition, NormalizeEventDefinitions, GenericCommandFactory, GenericCommandMap, resolveCommandHandler, createComponentBehaviorState, bindFluentMethods } from './redemeineComponent';
 import { bindContext } from './bindContext';
+import { applyEvent } from './applyEvent';
+import { defaultNamingStrategy } from './naming';
+import type { Merge } from './types/Merge';
+import type { AllKeys } from './types/AllKeys';
 
 type AggregateSelectorUtils = { bindContext: typeof bindContext };
