+++ packages/aggregate/src/index.ts
@@ -2,3 +2,4 @@ export * from './createAggregate';
 export * from './createEntity';
 export * from './createMixin';
 export * from './bindContext';
+export * from './naming';
