+++ packages/aggregate/src/redemeineComponent.ts
@@ -1,8 +1,8 @@
 import { CommandContext, CommandIntents, CommandResult, Event, PackedCommandWithMeta, PluginExtensions, SelectorsMap, ShorthandCommandWithMeta, ReadonlyDeep } from '@redemeine/kernel';
 import { createCommandContextProxy } from './proxies/createCommandContextProxy';
-import type { Merge } from '@redemeine/kernel';
-import type { AllKeys } from '@redemeine/kernel';
-import type { ReplaceFirstArg } from '@redemeine/kernel';
+import type { Merge } from './types/Merge';
+import type { AllKeys } from './types/AllKeys';
+import type { ReplaceFirstArg } from './types/ReplaceFirstArg';
 
 export type GenericSelectors = Record<string, unknown>;
 export type GenericCommandMap = Record<string, RedemeineCommandDefinition<any, Record<string, unknown>, {}>>;
