+++ packages/aggregate/src/types/NestedKeysOf.ts
