+++ packages/aggregate/src/types/NestedPairOf.ts
