+++ packages/aggregate/src/types/PickType.ts
