+++ packages/aggregate/src/types/ReplaceFirstArg.ts
