+++ packages/kernel/src/index.ts
@@ -3,13 +3,5 @@ export * from './createCommand';
 export * from './createEvent';
 export * from './identity';
 export * from './Contract';
-export * from './utils/naming';
-export * from './utils/applyEvent';
 export * from './utils/readonlyProxy';
-export type { AllKeys } from './utils/types/AllKeys';
-export type { Merge } from './utils/types/Merge';
-export type { NestedKeysOf } from './utils/types/NestedKeysOf';
-export type { NestedPairsOf } from './utils/types/NestedPairOf';
-export type { PickType } from './utils/types/PickType';
 export type { ReadonlyDeep } from './utils/types/ReadonlyDeep';
-export type { ReplaceFirstArg } from './utils/types/ReplaceFirstArg';
